backend/requirements.txt
# Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Auth
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Storage & Media
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
Pillow==11.1.0

# Cache
django-redis==5.4.0
redis==5.2.1

# Payments
stripe==11.4.1

# Utilities
python-dotenv==1.0.1
django-filter==24.3
whitenoise==6.8.2
gunicorn==23.0.0
celery==5.4.0
setuptools>=70.0.0